Finished isSolved method.
I did the isSolved method and commented everything I did. The idea to check if there is a repetition in a array is to use an hash table to keep on track which numbers appeared on it and then check if there is any position at the hash table where its value is not 1.